A* pathing class
Set up basic A*,

so the walking enemy would just have to declare a PathFinder, FindPath(), and continuously move toward path[0] until it reaches it, at which point we path.RemoveAt(0) and keep going.

I suspect the GetManhattanDistance thing to be woefully insufficient when dead-ends and weird tower mazes become involved, but still need to test.

Also, my GridTileScript doesn't seem to be setting the adjacent tiles all correctly, so that'd need to be fixed as well