LevelManager Script, spawnPoints, and waves
Dropped the TempGridScript, added LevelManager, which controls the current wave being sent and eventually the overall state of the game.

Also set up spawnpoint prefabs (red square gizmo in editor view), which you can place and adjust which enemies get sent from it with which wave.

Also goalTilePrefab (blue square gizmo in editor view), which then needs to be given to any instances of SpawnPoint so they know what endTile to aim for with their paths.

Paths are modifiable by changing a tile's walkable value or moving around the spawnpoint or goal tile. I changed the walkable value for a few tiles, shown by the gizmos being slightly different colors.