New Gravity WIP
changing gravity to instead rely on raycasting down from player to surface, and just setting the gravity dir and player's up transform based on the normal of the raycast hit. Will set up tag for specific surfaces with gravity changes for this raycast to look for.

Also need to set up airborne gravity changes, as that will need to have its own system...

Camera stuff, realized I need to add roll (we have pitch and yaw), which is equivalent to a camera's dutch, which cinemachines have, so I'm planning to switch back to cinemachines.