Gravity and pulse tower
Neither are fully functional as of yet. Gravity a little wonky atm (will fix before presentation), and pulse towers don't hit enemies.

Current main issue with gravity is rotation of player body when moving in third person camera. It throws things off once gravity changes. Planning to remove it if I can figure out strafing in time for presentation. Also player hovers above the ground sometimes, though that might be a result of the body rotations as well, because it wasn't an issue before that I can recall.

Will also try to add in Jorge's level before the presentation, ideally.