Enemy damage indicators, Jorge's terrain
Grabbed Jorge's scenes and terrain from his branch

Added damage indicators on enemies
TakeDamage in EnemyBehavior now instantiates a damageIndicator prefab (which is just a rigidbody and canvas, basically), which displays the damage value as text and flies away, lasting for 2s. Also uses the FaceCamera script from the healthbars.