tower fixes
added enemy.OnEnemyReachedGoal += RemoveEnemy in TowerBehavior. Issue seemed to be that we were only invoking RemoveEnemy for OnEnemyDeath, not for reaching the goal. So it didn't get removed first, then the enemy would be destroyed, and the list would be left with a null

Also added isAlive = true to other enemy behavior scripts' Start() functions (boolean is false by default, and was only set in Start() of EnemyBehavior)

Also dragged remainingEnemies scriptableObject into the inspectors of the enemy prefabs. that list works now for adding and removing properly, but still says type mismatch in the fields, and throws an error when you stop running