ingame HUD fixed
I broke it earlier; ingame HUD disappeared. Apparently the rectTransform of the InGameHUD prefab (when looked at through the UI Manager prefab) had scale and dimensions all at 0 for some reason idk. Fixed those to match the TowerUI (Which was showing up when interacting with a tower, and had 1920x1080 width and height and scale of 1, etc)