status redesign WIP
Putting handler of status list in its own MonoBehavior script: StatusEffects, to be attached to enemy prefabs (and probably tower prefabs and player)

holds all status effects on object in an array of Lists, with each list holding all instances of a given debuff type (so a list of all burns on the enemy, a list of all chills, etc). It increments the timeElapsed for all of those and removes them when their duration expires. For the first instance of each debuff type, the Effect() function is called and that List is passed to be processed (take the sum or multiplicative sum of all of that debuff type, and apply the total effect to the object it's on)